+++ f/frizzle/kobo/tests/conftest.py
@@ -2,16 +2,10 @@ import re
 from dataclasses import dataclass
 
 import pytest
-import responses
-import testing.postgresql
 
 from f.frizzle.kobo.tests.assets import server_responses
 
-
-@pytest.fixture
-def mocked_responses():
-    with responses.RequestsMock() as rsps:
-        yield rsps
+pytest_plugins = ["f.frizzle.tests.conftest"]
 
 
 @pytest.fixture
@@ -45,12 +39,3 @@ def koboserver(mocked_responses):
     )
 
     return KoboServer(dict(server_url=server_url, api_key="Callooh!Callay!"), form_id)
-
-
-@pytest.fixture
-def pg_database():
-    db = testing.postgresql.Postgresql(port=7654)
-    dsn = db.dsn()
-    dsn["dbname"] = dsn.pop("database")
-    yield dsn
-    db.stop
